add categories controller

+++ BLL/Services/GiftService.cs
@@ -14,12 +14,12 @@ namespace BLL.Services
     public class GiftService
     {
         private readonly IGiftRepository giftRepository;
-
         private readonly ITicketRepository ticketRepository;
 
-        public GiftService(IGiftRepository giftRepository)
+        public GiftService(IGiftRepository giftRepository, ITicketRepository ticketRepository)
         {
             this.giftRepository = giftRepository;
+            this.ticketRepository = ticketRepository;
         }
 
         public async Task<IEnumerable<GiftDto>> Get()
@@ -36,10 +36,36 @@ namespace BLL.Services
         }
         public async Task Update(int id, GiftDto gift)
         {
+            if (gift == null || gift.CategoryId == 0 || gift.DonorId == 0 || gift.GiftName == null)
+            {
+               throw new InvalidOperationException("Gift data cannot be null.");
+            }
+            if (gift.Price < 10 || gift.Price > 100)
+            {
+                throw new InvalidOperationException("Price must be between 10 and 100.");
+            }
+
+            var existingGift = await Get(id);
+            if (existingGift == null)
+            {
+                throw new KeyNotFoundException($"gift with id {id} not found.");
+            }
+
+            var existingGiftWithSameName = await TitleExists(gift.GiftName);
+            if (existingGiftWithSameName && existingGift.GiftName != gift.GiftName)
+            {
+                throw new DuplicateWaitObjectException("Gift with this name already exists.");
+            }
             await giftRepository.Update(id, gift);
+            
         }
         public async Task<bool> Delete(int id)
         {
+            var existingGift = await Get(id);
+            if (existingGift == null)
+            {
+                throw new KeyNotFoundException($"gift with id {id} not found.");
+            }
             return await giftRepository.Delete(id);
         }
         public async Task<IEnumerable<GiftDto>> Search(string giftName = null, string donorName = null, int? buyerCount = null)
